Split UI code into submodules
